Fixed eslint errors.
